Move keybinding accelerator generation to keybindings.ts
