+++ packages/core/src/browser/keybinding.ts
@@ -7,9 +7,10 @@
 
 import { injectable, inject, named } from 'inversify';
 import { CommandRegistry } from '../common/command';
-import { KeyCode, KeySequence } from './keys';
+import { Key, KeyCode, KeySequence } from './keys';
 import { ContributionProvider } from '../common/contribution-provider';
 import { ILogger } from "../common/logger";
+import { isOSX } from './../common/os';
 
 export enum KeybindingScope {
     DEFAULT,
@@ -36,6 +37,59 @@ export namespace Keybinding {
             context: binding.context
         };
         return JSON.stringify(copy);
+    }
+
+    /* Return a user visble representation of a keybinding.  */
+    export function acceleratorFor(keybinding: Keybinding) {
+        const keyCodesString = keybinding.keybinding.split(" ");
+        const result: string[] = [];
+        for (const keyCodeString of keyCodesString) {
+            let keyCodeResult = "";
+            const keyCode = KeyCode.parse(keyCodeString);
+            let previous = false;
+            const separator = " ";
+
+            if (keyCode.meta && isOSX) {
+                if (isOSX) {
+                    keyCodeResult += "Cmd";
+                    previous = true;
+                }
+            }
+
+            if (keyCode.ctrl) {
+                if (previous) {
+                    keyCodeResult += separator;
+                }
+                keyCodeResult += "Ctrl";
+                previous = true;
+            }
+
+            if (keyCode.alt) {
+                if (previous) {
+                    keyCodeResult += separator;
+                }
+                keyCodeResult += "Alt";
+                previous = true;
+            }
+
+            if (keyCode.shift) {
+                if (previous) {
+                    keyCodeResult += separator;
+                }
+                keyCodeResult += "Shift";
+                previous = true;
+            }
+
+            if (keyCode.key) {
+
+                if (previous) {
+                    keyCodeResult += separator;
+                }
+                keyCodeResult += Key.getEasyKey(keyCode.key).easyString.toUpperCase();
+            }
+            result.push(keyCodeResult);
+        }
+        return result;
     }
 }
